Download NLA data from USEPA

Retrieves NLA flat files.

nla_get(year, use_rappdirs = FALSE, local_path = tempdir(), compress = "xz")

Arguments

year
numeric choice of 2007 or 2012.

use_rappdirs
logical write files to operating system data directories at the location returned by user_data_dir.

local_path
folder path to raw downloads from `nla_get`

compress
a logical specifying whether saving to a named file is to use "gzip" compression, or one of "gzip", "bzip2" or "xz" to indicate the type of compression to be used. Ignored if file is a connection.

Examples

# \donttest{
nla_get(2012)

#> NLA 2012 compiled to /tmp/RtmpRSGT9T/data_2012.rds

# }